Move PuzzlePlugins out of :scrambles module to spare Kotlin dependencies
